# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAText - OpenAPI 3.0
  description: |-
    This is the OpenAPI specification for the WASAText web-app.
  version: 1.0.0
servers:
  - url: http://localhost:3000
...
paths:
  /session:
    post:
      tags: ["login"]
      summary: Logs in the user
      description: |-
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      security: []
      requestBody:
        description: |-
            User details
        content:
          application/json:
            schema:
              description: |-
                  This schema represents the requestBody of the "doLogin" operation.
              type: object
              properties:
                name:
                  description: |-
                      This propery represents the username of the user who wants to log in.
                  type: string
                  example: Maria
                  pattern: '^[A-Za-z0-9]{3,16}$'
                  minLength: 3
                  maxLength: 16
        required: true
      responses:
        "201":
          description: |-
              User log-in action successful
          content:
            application/json:
              schema:
                description: |-
                    This schema represents the 201 response to the doLogin operation.
                type: object
                properties:
                  identifier:
                    description: |-
                        The user identifier.
                    # change here if you decide to use an integer
                    # or any other type of identifier
                    type: string
                    example: "abcdef012345"
        "400":
          description: |-
              Invalid username
  /conversations:
    get:
      tags: ["conversations"]
      operationId: getMyConversations
      summary: Returns the conversations of the user
      description: |-
        Returns the lists of all the conversations of the user
      responses:
        "200":
          description: |-
            It is an array of integers
            containing the user's conversations' id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversations"
        "401":
          description: |-
            Unauthorized: missing or invalid bearer token
    post:
      tags: ["conversations"]
      operationId: createConversation
      summary: Create a new private conversation
      description: |-
        Creates a new private (one-to-one) conversation with the specified user
        and sends the first message at the same time.
        If a private conversation between the two users already exists, returns 409.
        The authenticated user is identified by the Bearer token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: |-
                  Request body for createConversation.
              type: object
              required:
                - targetUsername
                - content
              properties:
                targetUsername:
                  type: string
                  example: "Luigi"
                  minLength: 3
                  maxLength: 16
                  pattern: '^[A-Za-z0-9]{3,16}$'
                  description: |-
                      Username of the user to start a conversation with.
                content:
                  type: string
                  example: "Ciao"
                  pattern: '^.*?$'
                  minLength: 1
                  maxLength: 100000
                  description: |-
                      Text content of the first message.
      responses:
        "201":
          description: Conversation created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "400":
          description: Bad request - Invalid username or message format
        "401":
          description: Unauthorized
        "404":
          description: Target user not found
          content:
            application/json:
              schema:
                description: |-
                  This schema contains the 404 error related message
                type: object
                properties:
                  error:
                    description: |-
                      The 404 error related message (user not found)
                    type: string
                    example: "User 'Luigi' not found"
        "409":
          description: Conflict - A private conversation with this user already exists
          content:
            application/json:
              schema:
                description: |-
                  This schema contains the 409 error related message
                type: object
                properties:
                  error:
                    description: |-
                      The 409 error related message
                    type: string
                    example: "A conversation with 'Luigi' already exists"
  /conversations/{conversationId}:
    get:
      tags: ["conversations"]
      operationId: getConversation
      summary: Returns the specified conversation
      description: |-
        Returns the specified conversation among all the user's ones
      parameters:
        - name: conversationId
          in: path
          description: The id of the conversation
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Conversation returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - Not a participant in this conversation
        "404":
          description: Conversation not found
...
  /conversations/{conversationId}/read:
    parameters:
      - name: conversationId
        in: path
        description: The id of the conversation
        schema:
          type: integer
        required: true
    put:
      tags: ["messages"]
      operationId: markMessagesRead
      summary: Mark all messages in a conversation as read
      description: |-
        Marks all unread messages in the specified conversation as read
        for the authenticated user.
        Should be called when the user opens a conversation.
        The user is identified by the Bearer token.
      responses:
        "204":
          description: Messages marked as read successfully (no content)
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - Not a participant in this conversation
        "404":
          description: Conversation not found
  /conversations/{conversationId}/messages/{messageId}:
    parameters:
      - name: conversationId
        in: path
        description: |-
            The id of the conversation
        schema:
          type: integer
        required: true
      - name: messageId
        in: path
        description: |-
            The id of the message
        schema:
          type: integer
        required: true
    delete:
      tags: ["messages"]
      operationId: deleteMessage
      summary: The user deletes a previously sent message.
      description: |-
          The user deletes the specified message of the specified conversation.
      responses:
        "204":
          description: Message deleted successfully
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - Cannot delete messages from other users
        "404":
          description: Message not found
...
  /conversations/{conversationId}/messages/{messageId}/forward:
    post:
      tags: ["messages"]
      operationId: forwardMessage
      summary: Forward a message to another conversation
      description: |-
        Forwards the specified message to another conversation.
        The message content (text or photo) is copied to the target conversation.
        The user must be a participant in both the source and target conversations.
        The user is identified by the Bearer token.
      parameters:
        - name: conversationId
          in: path
          description: The id of the source conversation
          required: true
          schema:
            type: integer
            minimum: 1
        - name: messageId
          in: path
          description: |-
            The id of the message to forward
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationId"
      responses:
        "201":
          description: Message forwarded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageForwarded"
        "400":
          description: Bad request - Invalid conversation or message ID
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - Not a participant in source or target conversation
        "404":
          description: Source message, source conversation, or target conversation not found
  #USER PROFILE
...
  /groups/{groupId}/members/me:
    delete:
      tags: ["groups"]
      operationId: leaveGroup
      summary: Leave a group
      description: |-
        Removes the authenticated user from the specified group.
        The user is identified by the Bearer token.
        Users can leave groups at any time.
      parameters:
        - name: groupId
          in: path
          description: The id of the group to leave
          required: true
          schema:
            type: integer
            minimum: 1
            example: 42
      responses:
        "204":
          description: Successfully left the group (no content)
        "401":
          description: Unauthorized
        "404":
          description: Group not found or user is not a member
          content:
            application/json:
              schema:
                description: |-
                    This schema represents an error related to the 404 response of the "leaveGroup" operationId.
                type: object
                properties:
                  error:
                    description: |-
                        The string describing the error.
                    type: string
                    example: "Group not found or you are not a member"
#COMPONENTS
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |-
        User identifier returned from login.
        Use the identifier as the bearer token.
  schemas:
    Conversations:
      description: |-
          This schema represents a list of conversations
      type: array
      minItems: 0
      maxItems: 100000
      items:
        type: object
        properties:
          id:
            type: integer
            minimum: 1
            example: 42
            description: |-
                Conversation identifier
          type:
            type: string
            enum: [direct, group]
            example: direct
            description: |-
                Type of conversation (direct chat or group)
          name:
            type: string
            example: "Marco"
            description: |-
                Username (for direct) or group name (for group).
          photoUrl:
            type: string
            format: uri
            example: "https://example.com/photos/user123.jpg"
            description: |-
                Profile photo URL (user or group)
          lastMessage:
            type: object
            description: |-
                Preview of the last message
            properties:
              preview:
                type: string
                example: "Ciao, come stai?"
                maxLength: 100
                description: |-
                    Text snippet (photo url if photo message)
              timestamp:
                type: string
                format: date-time
                example: "2025-12-13T17:37:43Z"
                description: |-
                    Timestamp of last message
              isPhoto:
                type: boolean
                example: false
                description: |-
                    True if last message is a photo
          unreadCount:
            type: integer
            minimum: 0
            example: 3
            description: |-
                Number of unread messages in this conversation
    Conversation:
      description: |-
          This schema represents a conversation.
      type: object 
      properties:
        id:
          type: integer
          minimum: 1
          description: |-
              The conversation identifier
        type:
          type: string
          enum: [direct, group]
          example: direct
          description: |-
              Type of conversation
        name:
          type: string
          example: "Marco"
          description: |-
              Username or group name
        photoUrl:
          type: string
          format: uri
          example: "https://example.com/photos/user123.jpg"
          description: |-
              Profile photo URL
        participants:
          type: array
          description: |-
              List of participant usernames (for groups)
          minItems: 1
          maxItems: 10000000
          items:
            description: |-
                Each item of the array represents a single partecipant of the conversation.
            type: string
            example: "Maria"
        messages:
          type: array
          description: |-
              All messages in reverse chronological order
          minItems: 0
          maxItems: 100000000
          items:
            type: object
            properties:
              id:
                description: |-
                    The id of the message
                type: integer
                minimum: 1
                example: 18284
              senderId:
                type: string
                example: "user123"
                description: |-
                    ID of the message sender
              senderUsername:
                type: string
                example: "Maria"
                description: |-
                    Username of sender
              content:
                type: string
                example: "Ciao, come stai?"
                description: |-
                    Text content (empty if photo message)
              photoUrl:
                type: string
                format: uri
                example: "https://example.com/photos/msg123.jpg"
                description: |-
                    Photo URL (present only for photo messages)
              timestamp:
                description: |-
                    The timesatmp of the message.
                type: string
                format: date-time
                example: "2025-12-13T17:37:43Z"
              status:
                type: string
                enum: [sent, received, read]
                example: "received"
                description: |-
                  Message status (checkmarks - for sent messages):
                    - sent: message sent (no checkmark)
                    - received: one checkmark (delivered to all recipients)
                    - read: two checkmarks (opened by all recipients)
              replyTo:
                type: object
                description: |-
                    Present only if this message is a reply
                properties:
                  messageId:
                    description: |-
                        The id of the message to be replied to.
                    type: integer
                    example: 18200
                  content:
                    description: |-
                        The content of the message to be replied to.
                    type: string
                    example: "Ci sei?"
                  senderUsername:
                    description: |-
                        The username of user who sent the message to be replied to
                    type: string
                    example: "Luigi"
              reactions:
                type: array
                description: |-
                    Comments/reactions on this message
                minItems: 0
                maxItems: 100000
                items:
                  type: object
                  properties:
                    username:
                      description: |-
                          The username of the user who reacted.
                      type: string
                      example: "Marco"
                    emoticon:
                      type: string
                      example: "👍"
                      description: |-
                          Reaction emoticon
                    timestamp:
                      description: |-
                          The timestamp of the reaction.
                      type: string
                      format: date-time
                      example: "2025-12-13T17:38:00Z"
              isForwarded:
                type: boolean
                example: false
                description: |-
                    True if this message was forwarded from another conversation
...
    MessageForwarded:
      description: |-
          This schema represents a message forwarded from a conversation to another.
      type: object
      properties:
        id:
          description: |-
              The id of the message.
          type: integer
          minimum: 1
          example: 18999
        checkmarks:
          description: |-
              The checkmarks of the message (1 if every user of the conversation received it, 2 if every user of the conversation read it, 0 if at least a user did not receive it).
          type: integer
          minimum: 0
          maximum: 2
          example: 0
        timestamp:
          description: |-
              The timestamp of the message
          type: string
          format: date-time
        content:
          description: |-
              The content of the message, if it is a string.
          type: string
          example: "Ciao, come stai?"
        photoUrl:
          type: string
          format: url
          description: |-
              Present only if message contains a photo
        targetConversationId:
          description: |-
              The id of the conversation to which the message belongs.
          type: integer
          example: 99
    ConversationId:
      description: |-
          This schema represents the id of a particular conversation(not group, so it is supposed to be a conversation between 2 users only)
      type: object
      required:
        - targetConversationId
      properties:
        targetConversationId:
          type: integer
          minimum: 1
          example: 99
          description: |-
            The id of the conversation to which the message is dirrected.
...
security:
  - bearerAuth: []